Feat fix builds after godot-cpp downgrade to 4.2 (#41)
* fix build errors

+++ src/eosg_multiplayer_peer.cpp
@@ -1244,7 +1244,7 @@ bool EOSGMultiplayerPeer::EOSGSocket::_socket_id_is_valid(const String &socket_i
     int alpha_lowercase_range_max = 122;
 
     CharString str = socket_id.ascii();
-    for (int64_t i = 0; i < str.length(); i++) {
+    for (int i = 0; i < str.length(); i++) {
         if (str[i] >= numeric_range_min && str[i] <= numeric_range_max) {
             continue;
         } else if (str[i] >= alpha_capitalized_range_min && str[i] <= alpha_capitalized_range_max) {
